Refactor: 통계 페이지들의 스타일 분리
